+++ test/crossfile/definition.lua
@@ -770,7 +770,7 @@ TEST {
     },
 }
 
-config.set('Lua.IntelliSense.traceGlobalInject', true)
+config.set('Lua.IntelliSense.traceFieldInject', true)
 TEST {
     {
         path = 'a.lua',
@@ -791,7 +791,7 @@ print(b.<?test?>)
         ]]
     }
 }
-config.set('Lua.IntelliSense.traceGlobalInject', false)
+config.set('Lua.IntelliSense.traceFieldInject', false)
 
 TEST {
     {